# openapi.yaml
openapi: "3.1.0"
info:
  title: Orchestrator Platform API
  description: |
    Multi-Model AI Agent Orchestration Platform API.
    Provides a unified gateway to multiple LLM providers (OpenAI, Anthropic, Google),
    isolated code execution sandboxes, and a workflow orchestration engine that
    decomposes objectives into DAGs of parallel tasks.
  version: "0.1.0"
  contact:
    name: Orchestrator Platform
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key authentication. Pass your key as `Bearer sk-...`
  schemas:
...
    SandboxConfig:
      type: object
      required: [language]
      properties:
        language:
          type: string
          enum: [python, javascript, sql]
        timeout_seconds:
          type: integer
          maximum: 3600
        packages:
          type: array
          items:
            type: string
        files:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              content_base64:
                type: string
    SandboxSession:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [creating, ready, executing, terminated, error]
        language:
          type: string
        created_at:
          type: string
    ExecutionResult:
      type: object
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exit_code:
          type: integer
        execution_time_ms:
          type: number
        files_modified:
          type: array
          items:
            type: string

    # ── Workflows ──
...
paths:
...
  /v1/billing/usage:
    get:
      summary: Get usage details
      tags: [Billing]
      parameters:
        - name: start
          in: query
          schema:
            type: string
        - name: end
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Usage details
  /v1/billing/top-up:
    post:
      summary: Add credits
      tags: [Billing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  minimum: 0
      responses:
        "200":
          description: Credits added
          content:
            application/json:
              schema:
                type: object
                properties:
                  credits_balance:
                    type: number
                  amount_added:
                    type: number
  /v1/billing/transactions:
    get:
      summary: List credit transactions
      tags: [Billing]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      type: object

  # ── Sandbox ──
  /v1/sandbox/sessions:
    post:
      summary: Create a sandbox session
      tags: [Sandbox]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SandboxConfig'
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxSession'
  /v1/sandbox/sessions/{id}:
    get:
      summary: Get session info
      tags: [Sandbox]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxSession'
    delete:
      summary: Terminate a session
      tags: [Sandbox]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session terminated
  /v1/sandbox/sessions/{id}/execute:
    post:
      summary: Execute code in a session
      tags: [Sandbox]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                timeout_seconds:
                  type: integer
                  maximum: 300
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResult'
  /v1/sandbox/sessions/{id}/files:
    get:
      summary: List files in workspace
      tags: [Sandbox]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: directory
          in: query
          schema:
            type: string
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: string
...